requirements.txt
asgiref==3.2.3
attrs==19.3.0
dj-database-url==0.5.0
Django==2.2.5
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
importlib-metadata==1.5.0
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
psycopg2==2.7.4
psycopg2-binary==2.8.3
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
wcwidth==0.1.8
whitenoise==5.0.1
zipp==2.1.0
